+++ src/main/java/dev/jpp/listapi/service/ISimpleCrudService.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface ISimpleCrudService<T> {
 	public abstract T add(T model);
 	public abstract T update(T model);
-	public abstract void remove(int id);
+	public abstract void remove(Long id);
 	public abstract List<T> findAll();
-	public abstract T findById(int id);
+	public abstract T findById(Long id);
 }
